+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Reevaluate `window.fetch` each time `HttpLink` uses it, if not configured using `options.fetch`. This change enables a variety of strategies for instrumenting `window.fetch`, without requiring those strategies to run before `@apollo/client/link/http` is first imported. <br/>
   [@benjamn](https://github.com/benjamn) in [#8603](https://github.com/apollographql/apollo-client/pull/8603)
 
+- Clarify mutation `fetchPolicy` options (`"network-only"` or `"no-cache"`) using `MutationFetchPolicy` union type. <br/>
+  [@benjamn](https://github.com/benjamn) in [#8602](https://github.com/apollographql/apollo-client/pull/8602)
+
 ### Bug Fixes
 
 - Restore full `@apollo/client/apollo-client.cjs.js` CommonJS bundle for older bundlers.